Move ConsoleTask to pants/task.
It doesn't belong in backend/core.

There are forwarding aliases in the original location, so
custom tasks referencing them won't break (but will display
a deprecation warning).

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/92390542

Reviewed at https://rbcommons.com/s/twitter/r/3157/
