+++ src/python/pants/backend/core/tasks/filemap.py
@@ -5,7 +5,7 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.core.tasks.console_task import ConsoleTask
+from pants.task.console_task import ConsoleTask
 
 
 class Filemap(ConsoleTask):
