+++ src/python/pants/backend/core/tasks/list_goals.py
@@ -5,8 +5,8 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.goal.goal import Goal
+from pants.task.console_task import ConsoleTask
 
 
 class ListGoals(ConsoleTask):
