+++ src/python/pants/backend/core/tasks/list_owners.py
@@ -5,9 +5,9 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.base.exceptions import TaskError
 from pants.build_graph.source_mapper import LazySourceMapper
+from pants.task.console_task import ConsoleTask
 
 
 class ListOwners(ConsoleTask):
