+++ src/python/pants/backend/core/tasks/listtargets.py
@@ -5,8 +5,8 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.base.exceptions import TaskError
+from pants.task.console_task import ConsoleTask
 
 
 class ListTargets(ConsoleTask):
