+++ src/python/pants/backend/core/tasks/sorttargets.py
@@ -7,8 +7,8 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from collections import defaultdict
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.build_graph.build_graph import sort_targets
+from pants.task.console_task import ConsoleTask
 
 
 class SortTargets(ConsoleTask):
