+++ src/python/pants/backend/jvm/tasks/jvm_platform_analysis.py
@@ -11,11 +11,11 @@ from hashlib import sha1
 
 from colors import red
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.backend.jvm.targets.jvm_target import JvmTarget
 from pants.base.exceptions import TaskError
 from pants.base.fingerprint_strategy import FingerprintStrategy
 from pants.build_graph.build_graph import CycleException, sort_targets
+from pants.task.console_task import ConsoleTask
 from pants.task.task import Task
 from pants.util.memo import memoized_property
