+++ src/python/pants/backend/project_info/tasks/depmap.py
@@ -5,9 +5,9 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.backend.jvm.targets.jar_dependency import JarDependency
 from pants.base.exceptions import TaskError
+from pants.task.console_task import ConsoleTask
 
 
 class Depmap(ConsoleTask):
