+++ tests/python/pants_test/tasks/task_test_base.py
@@ -10,9 +10,9 @@ import subprocess
 from contextlib import closing
 from StringIO import StringIO
 
-from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.goal.goal import Goal
 from pants.ivy.bootstrapper import Bootstrapper
+from pants.task.console_task import ConsoleTask
 from pants.util.contextutil import temporary_dir
 from pants_test.base_test import BaseTest
