+++ tests/python/pants_test/tasks/test_console_task.py
@@ -9,7 +9,7 @@ import os
 import threading
 from Queue import Empty, Queue
 
-from pants.backend.core.tasks.console_task import ConsoleTask
+from pants.task.console_task import ConsoleTask
 from pants_test.tasks.task_test_base import TaskTestBase
